+++ templates/element/footer.php
@@ -25,7 +25,7 @@ if (getConfig('default.contact_us')) {
     $links[] = $this->Html->link(__d('me_cms', 'Contact us'), ['_name' => 'contactUs']);
 }
 
-echo $this->Html->ul($links);
+echo $this->Html->ul($links, ['class' => 'list-inline'], ['class' => 'list-inline-item border-end mx-0 px-2']);
 
-echo $this->Html->para('', __d('me_cms', 'Powered by {0}. Copyright {1}', 'MeCms', date('Y')));
-echo $this->Html->para('', __d('me_cms', 'Developed by {0}', $this->Html->link('Mirko Pagliai', '//github.com/mirko-pagliai', ['target' => '_blank'])));
+echo $this->Html->para('mb-0', __d('me_cms', 'Powered by {0}. Copyright {1}', 'MeCms', date('Y')));
+echo $this->Html->para('mb-0', __d('me_cms', 'Developed by {0}', $this->Html->link('Mirko Pagliai', 'https://github.com/mirko-pagliai', ['target' => '_blank'])));