+++ tests/TestCase/Template/Element/TopbarTest.php
@@ -32,11 +32,11 @@ class TopbarTest extends TestCase
     {
         $View = new AppView();
         $result = $View->element('MeCms.topbar');
-        $this->assertStringContainsString('<a href="/" class="nav-link" title="Home"><i class="fas fa-home"> </i> Home</a>', $result);
+        $this->assertStringContainsString('<a href="/" class="nav-link" title="Home">', $result);
 
         $this->loadPlugins(['TestPlugin' => []]);
         $result = $View->element('MeCms.topbar');
-        $this->assertStringContainsString('<a href="/" class="nav-link" title="Home"><i class="fas fa-home"> </i> Home</a>', $result);
+        $this->assertStringContainsString('<a href="/" class="nav-link" title="Home">', $result);
         $this->assertStringContainsString('<a href="http://localhost" title="A link from plugin">A link from plugin</a>', $result);
     }