+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2.30 branch
 ### 2.30.11
 * the topbar has been made permanently compatible with the new bootstrap;
+* the default layout can automatically establish whether to use a sidebar and whether to directly use the 
+  css files (`layout.css` and `contents.css`) from the application webroot;
 * fixed little bugs for posts preview;
 * removed a lot of useless css code and optimized the layout and its elements.